+++ R/recover.R
@@ -11,7 +11,7 @@ recover <- function(result_folder, maxResources = 90,
   }
 
   ssEnv <- init_env( result_folder =  result_folder, maxResources =  maxResources, parallel_strategy  =  parallel_strategy, start_fresh = FALSE, ...)
-  log_event("BANNER: ", format(Sys.time(), "%a %b %d %X %Y"), " SemSeeker will try to recover the analysys for project \n in ", ssEnv$result_folderData)
+  log_event("BANNER: ", format(Sys.time(), "%a %b %d %X %Y"), " SEMseeker will try to recover the analysys for project \n in ", ssEnv$result_folderData)
 
   study_summary <-   study_summary_get()